Init setpoint to current state before closed-loop
When going from open-loop to closed-loop, the setpoint (not the goal)
should be re-set to the current state.  This re-syncs the setpoint and
the physical elevator, which may be far from each other because the
elevator was not under closed-loop control before.

Otherwise the setpoint will ramp from where it was previously to the
new goal, causing the physical elevator to try to jump to the old
setpoint before it converges on the new one.