Update controller all the time
It's simpler to update the controller only while the elevator is
running closed loop, but this causes a transient when going from
open- to closed-loop.

It's also tempting to put the call to updateController() call in
periodic(), but periodic is called before the active command runs,
so this would mean an extra loop delay on all setpoint updates.

I'm still not entirely happy with the way this code reads, but it
runs great.